Move coroutine layout logic to `rustc_abi`
